Unified BraveVpnServiceDesktop into BraveVpnService
No functional changes.
It's preparation for sharing the purchased state managing logic
on Desktop and Android.
