+++ browser/brave_vpn/brave_vpn_service_factory.cc
@@ -16,16 +16,7 @@
 #include "content/public/browser/browser_context.h"
 #include "content/public/browser/storage_partition.h"
 
-// TODO(bsclifton) or TODO(shong):
-// We should be able to consolidate this integration into one implementation
-// which we can share between Android and Desktop.
-//
-// As seen below, Desktop returns BraveVpnServiceDesktop and Android
-// returns BraveVpnService.
-//
-// See https://github.com/brave/brave-browser/issues/20374 for more info.
 #if BUILDFLAG(IS_WIN) || BUILDFLAG(IS_MAC)
-#include "brave/components/brave_vpn/brave_vpn_service_desktop.h"
 #include "brave/components/brave_vpn/brave_vpn_utils.h"
 #endif
 
@@ -35,24 +26,11 @@ BraveVpnServiceFactory* BraveVpnServiceFactory::GetInstance() {
 }
 
 // static
-#if BUILDFLAG(IS_WIN) || BUILDFLAG(IS_MAC)
-BraveVpnServiceDesktop* BraveVpnServiceFactory::GetForProfile(
-    Profile* profile) {
-  return static_cast<BraveVpnServiceDesktop*>(
-      GetInstance()->GetServiceForBrowserContext(profile, true));
-}
-#elif BUILDFLAG(IS_ANDROID)
 BraveVpnService* BraveVpnServiceFactory::GetForProfile(Profile* profile) {
   return static_cast<BraveVpnService*>(
       GetInstance()->GetServiceForBrowserContext(profile, true));
 }
-#endif
 
-// TODO(bsclifton) or TODO(shong):
-// BraveVpnServiceDesktop is currently only used on Desktop,
-// which is why there are only OS guards for Windows and macOS.
-// Consolidating the Android/Desktop behaviors is captured with:
-// https://github.com/brave/brave-browser/issues/20374
 BraveVpnServiceFactory::BraveVpnServiceFactory()
     : BrowserContextKeyedServiceFactory(
           "BraveVpnService",
@@ -75,15 +53,13 @@ KeyedService* BraveVpnServiceFactory::BuildServiceInstanceFor(
   auto shared_url_loader_factory =
       default_storage_partition->GetURLLoaderFactoryForBrowserProcess();
 
-// TODO(bsclifton) or TODO(shong):
-// see same notes above
 #if BUILDFLAG(IS_WIN) || BUILDFLAG(IS_MAC)
   auto callback = base::BindRepeating(
       [](content::BrowserContext* context) {
         return skus::SkusServiceFactory::GetForContext(context);
       },
       context);
-  return new BraveVpnServiceDesktop(
+  return new BraveVpnService(
       shared_url_loader_factory, user_prefs::UserPrefs::Get(context), callback);
 #elif BUILDFLAG(IS_ANDROID)
   return new BraveVpnService(shared_url_loader_factory);
