+++ browser/brave_vpn/brave_vpn_service_factory.h
@@ -12,25 +12,12 @@
 #include "build/build_config.h"
 #include "components/keyed_service/content/browser_context_keyed_service_factory.h"
 
-class Profile;
-
-#if BUILDFLAG(IS_WIN) || BUILDFLAG(IS_MAC)
-class BraveVpnServiceDesktop;
-#endif
-
-#if BUILDFLAG(IS_ANDROID)
 class BraveVpnService;
-#endif
+class Profile;
 
 class BraveVpnServiceFactory : public BrowserContextKeyedServiceFactory {
  public:
-#if BUILDFLAG(IS_WIN) || BUILDFLAG(IS_MAC)
-  static BraveVpnServiceDesktop* GetForProfile(Profile* profile);
-#endif
-
-#if BUILDFLAG(IS_ANDROID)
   static BraveVpnService* GetForProfile(Profile* profile);
-#endif
   static BraveVpnServiceFactory* GetInstance();
 
   BraveVpnServiceFactory(const BraveVpnServiceFactory&) = delete;
